// src/app/models/vtas/Cliente.ts
export class Cliente {
  actividadEmpresaId: number;
  actividadEmpresas: any;
  activo: boolean;
  agencias: any;
  clasificacion: number;
  contactos: any;
  direcciones: any;
  emails: any;
  esCliente: boolean;
  fch_Creacion: Date;
  fch_Modificacion: Date;
  giroEmpresaId: number;
  giroEmpresas: any;
  id: string;
  nombrecomercial: string;
  numeroEmpleados: number;
  otraAgencia: boolean;
  razonSocial: string;
  referenciados: any;
  rfc: string;
  tamanoEmpresaId: number;
  tamanoEmpresas: string;
  telefonos: any;
  tipoBaseId: number;
  tipoBases: any;
  tipoEmpresaId: number;
  tipoEmpresas: string;
  tipoEntidad: any;
  tipoentidadId: number;
}